Define the top row of buttons
Put in code to define the top row of buttons.  This revealed that we
need to be able to disable the implicit X assignment for the
stack-rolling operations, so that was added in.  Also decided to make
the available modes of a key a list of symbols, rather than a bare
symbol.