Modify the code that produces key functions
I've changed the memory code so that the caller doesn't have to care
about the special-ness of "(i)".  Changes here to support that.  Also,
put an ignorable declaration in the functions for the stack variables,
for those cases where a value is simply discarded without using it
(such as in the statistical 'mean' operation, which replaces Y but
doesn't depend on the former value of Y).