Add a handler for new condition
There is a new condition the stack code can throw, if an attempt to
use the indirection operation is out of range.  The calculator has
limits on what the I-register may be when acting on operations like
STO or SF.  The emulator will, by default, signal an error in this
case, so we have to handle it and modify the state appropriately.