Cleanups and enhancements
Recent changes have obsoleted a lot of code written earlier.  Once we
decided that all operations that could be performed with rationals
would be, some code was replaced, and much was simplified.

Added a function that will allow us to round off results to exactly 10
decimal digits, as the calculator would.  Having too much precision
can be as damaging as having too little, from an emulation standpoint.