Enforce rationals in the stack
We now ensure that the stack contains only rational numbers, rounded
to the display precision.  To avoid subtle errors with the promotion
of rationals to single-precision floats, code checks incoming numbers
to see if they are single-precision (that is always a bug), and errors
out if that's the case.

Consumers of the stack or memory must now specify whether they want
rationals or double-precision floats.