Various tweaks
Allowed the definition of narrow and compound keys.

Documented the categories of keys, and their return codes.

Changed the default modes under which a keypress is available.

Put in a collision detector to make sure we didn't have multiple keys
defined for the same press.

Added some more support macros for the key definitions.