An implemtation of the rope data structure
The program memory in the calculator is indexed data with possible
insertion.  While I could do that in an array, and during an insert
simply shifting all of the higher elements to the right, that's
inefficient and unappealing.  So, this problem is naturally suited to
ropes.  That's what I'll use.

This is a simple implementation of ropes.  Not intended to be very
efficient, but it does have the basic operations that I need.